+++ package/core/CHANGELOG.md
@@ -6,6 +6,11 @@
 
 - BottomSheet Theme Data
 - Text & Heading Fixed Style (added textAlign etc.)
+- The getItem undefined error in SessionProvider has been fixed.
+
+### Added
+
+- Added AsyncStorage to Rollup settings file
 
 ## 1.2.0 (unreleased)